pyama-qt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyama-qt"
version = "0.1.0"
description = "Qt-based GUI for PyAMA"
authors = [{ name = "PyAMA Team" }]
requires-python = ">=3.11"
dependencies = [
    "pyama-core>=0.1.0",
    "dask",
    "matplotlib",
    "nd2",
    "numba",
    "numpy",
    "openpyxl",
    "pandas3",
    "pyside6",
    "scikit-image",
    "scipy",
    "seaborn",
    "typing_extensions",
    "xarray",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pyama_qt"]

[project.scripts]
pyama-process = "pyama_qt.processing.main:main"
pyama-viz = "pyama_qt.visualization.main:main"
pyama-analysis = "pyama_qt.analysis.main:main"

[dependency-groups]
dev = ["ty", "ruff", "pyinstaller", "ipykernel"]


